refactor test_dataframe & test_series
